Add files via upload
Hello Dr. Hart, 

Attached, please see my attempt at the Codebook assignment. I produced a RMarkdown document and then made changes that made the knit fail, so all that is here is the Data and the code from the RMarkdown document. Additionally I did not prioritize the assignment high enough on my list of things to accomplish since we last met and it shows (I struggled to locate the left-right self-placement and survey weight variables as well as to split the date of interview to produce year and month variables. I have no excuses for this and I am very disappointed in myself. I sincerely appreciate your effort to push me in the right direction. 

Best regards,

Camryn